MEDIUM: quic: count quic_conn for global sslconns
Similar to the previous commit which check for maxconn before allocating
a QUIC connection, this patch checks for maxsslconn at the same step.
This is necessary as a QUIC connection cannot run without a SSL context.

This should be backported up to 2.6. It relies on the following patch :
  "BUG/MINOR: ssl: use a thread-safe sslconns increment"

(cherry picked from commit 4a89dba6d576677219ec897fc3ab99e8f40d3923)
[cf: The code differs from the 2.9. The patch was applied on quic_conn.c.
     qc_free_ssl_sock_ctx() and qc_alloc_ssl_sock_ctx() don't exist.
     qc_conn_alloc_ssl_ctx() and quic_conn_release() was fixed instead]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>