BUG/MINOR: stconn: Sanitize report for read activity
When a EOS or EOI is detected on the endpoint and when the event is reported
at the SC level, a read activity must be reported. It is not really a big
deal because these flags already inhibit any read timeout. But it is
consistent with the <lra> comment. In addition, no read activity is reported
on abort. It is up-down event and it is not an event unblocking the
reads. So there is no reason to report a read activity.

This patch must be backported to 2.8.

(cherry picked from commit c57af8ebcdd39bcef7c3edc2d1d6e11b6ef7f448)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>