BUG/MINOR: stconn/applet: Report send activity only if there was output data
For applets and connection, when a send attempt is performed, we must be
sure to not report a send activity if there was no output data at all before
the attempt.

It is not important for the <fsb> date itself but for the <lra> date for
non-independent stream.

This patch must be backported to 2.8.

(cherry picked from commit 7676a2cdf6ceb3fa68156f7c47ade5e23c4094fd)
[cf: ctx adjt]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>