MINOR: quic: Use the correct socket file descriptor for QUIC servers
The socket value used to connect to QUIC servers is stored into ->fd quic_conn
struct member and initialized by ->connect() QUIC protocol callback (proto_quic.c).
This is also the case for QUIC listeners when they do not own the socket (the
connection is the socket owner).

Modify the code section which call sendto() to emit QUIC datagrams in this two
mode to use the correct file descriptor for QUIC servers.