BUG/MINOR: server: reject enabled for dynamic server
Since first implementation, dynamic servers are created into maintenance
state. This has been done purposely to avoid immediate activation of a
newly inserted server.

However, this principle is incompatible with "enabled" keyword. If used
on "add server", the newly created instance will still yet be
deactivated as load-balancing algorithm is not informed of the new
server via srv_lb_propagate(). The new server could be unblocked by
toggling its state with "disable server" / "enable server" commands,
which will trigger srv_lb_propagate() invocation.

To avoid this unexpected state, simply reject "enabled" keyword for
dynamic servers. In the long-term, it could be possible to support it
but for now it's unsure if it is safe to use srv_lb_propagate() after
dynamic server insertion.

This should fix github issue #2497.

This should be backported up to 2.4.